+++ src/reuse/_comment.py
@@ -1,5 +1,6 @@
-# SPDX-FileCopyrightText: 2019 Free Software Foundation Europe e.V.
+# SPDX-FileCopyrightText: 2019-2020 Free Software Foundation Europe e.V.
 # SPDX-FileCopyrightText: 2019 Kirill Elagin
+# SPDX-FileCopyrightText: 2020 Dmitry Bogatov
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
@@ -255,6 +256,12 @@ class AspxCommentStyle(CommentStyle):
     MULTI_LINE = ("<%--", "", "--%>")
 
 
+class BibTexCommentStyle(CommentStyle):
+    """BibTex comment style."""
+
+    MULTI_LINE = ("@Comment{", "", "}")
+
+
 class CCommentStyle(CommentStyle):
     """C comment style."""
 
@@ -335,7 +342,7 @@ class TexCommentStyle(CommentStyle):
 
 
 #: A map of (common) file extensions against comment types.
-COMMENT_STYLE_MAP = {
+EXTENSION_COMMENT_STYLE_MAP = {
     ".adb": HaskellCommentStyle,
     ".ads": HaskellCommentStyle,
     ".ahk": LispCommentStyle,
@@ -349,6 +356,7 @@ COMMENT_STYLE_MAP = {
     ".axd": AspxCommentStyle,
     ".applescript": AppleScriptCommentStyle,
     ".bash": PythonCommentStyle,
+    ".bib": BibTexCommentStyle,
     ".c": CCommentStyle,
     ".cl": LispCommentStyle,
     ".clj": LispCommentStyle,
@@ -435,11 +443,26 @@ COMMENT_STYLE_MAP = {
     ".zsh": PythonCommentStyle,
 }
 
+FILENAME_COMMENT_STYLE_MAP = {
+    ".gitattributes": PythonCommentStyle,
+    ".gitignore": PythonCommentStyle,
+    ".gitmodules": PythonCommentStyle,
+    ".editorconfig": PythonCommentStyle,
+    ".pylintrc": PythonCommentStyle,
+    "Dockerfile": PythonCommentStyle,
+    "Makefile": PythonCommentStyle,
+    "Manifest.in": PythonCommentStyle,
+    "manifest": PythonCommentStyle,  # used by cdist
+    "requirements.txt": PythonCommentStyle,
+    "setup.cfg": PythonCommentStyle,
+}
+
 # IMPORTANT: !!! When adding a new style, also edit usage.rst !!!
 #: A map of human-friendly names against style classes.
 NAME_STYLE_MAP = {
     "applescript": AppleScriptCommentStyle,
     "aspx": AspxCommentStyle,
+    "bibtex": BibTexCommentStyle,
     "c": CCommentStyle,
     "css": CssCommentStyle,
     "haskell": HaskellCommentStyle,